LP64 #5 round 2: style__ReadAttr long operand, figure_NULLREF -1L, eq mark sentinel, call-site casts
FossilOrigin-Name: 88171bcfbf8a57ed317044d61074870c3aaa327a8bb7898f0fe90b33f0ba489b

+++ src/atk/srctext/srctext.c
@@ -192,8 +192,8 @@ struct srctext *self;
 	style_AddNewFontFace(global, fontdesc_Fixed);
 	style_AddFlag(global, style_ContinueIndent);
 	style_AddFlag(global, style_TabsCharacters);
-	style_SetNewLeftMargin(global, style_LeftEdge, 16, style_RawDots);
-	style_SetNewIndentation(global, style_LeftMargin, -16, style_RawDots);
+	style_SetNewLeftMargin(global, style_LeftEdge, (long)16, style_RawDots);
+	style_SetNewIndentation(global, style_LeftMargin, (long)-16, style_RawDots);
 	style_SetJustification(global, style_LeftJustified);
 	stylesheet_Add(ss, global);
     }
